Run LUT conversion on LUT device
The inputs to the LUTs are always put, if need be, on the device of the LUT itself. It prevents overflows on the GPU memory for large images.

To do so, the dataset (e.g.  `DatasetCityscapesSemantic`) must be initialized with the right device, i.e. "cpu" to prevent VRAM overflows.